+++ lib/locale/translations.dart
@@ -22,13 +22,20 @@ class Translations {
     return _localizedValues[key] ?? '** $key not found';
   }
 
-  static Future<Translations> load(Locale locale) async {
+  static Future<Translations> load(BuildContext context, Locale locale) async {
     Translations translations = new Translations(locale);
-    String jsonContent = await rootBundle.loadString(
-          "res/values/strings_${locale.languageCode}_${locale.countryCode}.arb");
-    if(jsonContent == null || jsonContent.length <= 0){
+    String jsonContent = '';
+    try {
       jsonContent = await rootBundle.loadString(
-          "res/values/strings_en_US.arb");
+          "res/values/strings_${locale.languageCode}_${locale
+              .countryCode}.arb");
+      if(jsonContent == null || jsonContent.length <= 0) {
+        print("current language json is empty, use default language json");
+        jsonContent = await rootBundle.loadString("res/values/strings_en_US.arb");
+      }
+    }on Error catch (e) {
+      print(e.toString());
+      jsonContent =  await rootBundle.loadString("res/values/strings_en_US.arb");
     }
     _localizedValues = json.decode(jsonContent);
     return translations;
@@ -41,13 +48,15 @@ class Translations {
 
 class TranslationsDelegate extends LocalizationsDelegate<Translations> {
 
-  const TranslationsDelegate();
+  final BuildContext context;
+
+  const TranslationsDelegate(this.context);
 
   @override
-  bool isSupported(Locale locale) => ['en', 'zh', 'es', 'cz', 'sk'].contains(locale.languageCode);
+  bool isSupported(Locale locale) => true;
 
   @override
-  Future<Translations> load(Locale locale) => Translations.load(locale);
+  Future<Translations> load(Locale locale) => Translations.load(context, locale);
 
   @override
   bool shouldReload(TranslationsDelegate old) => false;